allow import llamaindex/vector-store/[VECTOR-STORE-NAME] for all vector stores
