+++ packages/llamaindex/src/vector-store/AstraDBVectorStore.ts
@@ -1,0 +1,1 @@
+export * from "@llamaindex/astra";
