+++ packages/llamaindex/src/vector-store/AzureAISearchVectorStore.ts
@@ -1,0 +1,1 @@
+export * from "@llamaindex/azure";
