+++ packages/llamaindex/src/vector-store/ChromaVectorStore.ts
@@ -1,0 +1,1 @@
+export * from "@llamaindex/chroma";
