+++ packages/llamaindex/src/vector-store/MilvusVectorStore.ts
@@ -1,0 +1,1 @@
+export * from "@llamaindex/milvus";
