+++ packages/llamaindex/src/vector-store/MongoDBAtlasVectorStore.ts
@@ -1,0 +1,1 @@
+export * from "@llamaindex/mongodb";
