+++ packages/llamaindex/src/vector-store/PGVectorStore.ts
@@ -1,0 +1,7 @@
+export {
+  DEFAULT_DIMENSIONS,
+  PGVECTOR_SCHEMA,
+  PGVECTOR_TABLE,
+  PGVectorStore,
+  type PGVectorStoreConfig,
+} from "@llamaindex/postgres";
