+++ packages/llamaindex/src/vector-store/WeaviateVectorStore.ts
@@ -1,0 +1,1 @@
+export * from "@llamaindex/weaviate";
